Split binds_p2/binds_p3/binds_p4 at clean impl-block boundaries to clear max-lines
