+++ crates/almide-mir/src/lower/binds_p2.rs
@@ -609,1203 +609,4 @@ impl LowerCtx {
         }
         self.lower_bind_heap_fresh_opaque(var, ty, value)
     }
-
-    /// Extracted from `Self::lower_bind_heap_fresh` (second-round split, cog reduction):
-    /// the concat/interp/str-list/option-ctor quick wins, verbatim. `Ok(true)` means the
-    /// caller already bound `var` and should return immediately.
-    fn try_lower_bind_heap_fresh_quick(
-        &mut self,
-        var: VarId,
-        ty: &Ty,
-        value: &IrExpr,
-    ) -> Result<bool, LowerError> {
-        // `let s = a + b` — a string concat EXECUTES to a fresh owned String (via the
-        // self-host __str_concat), held by the binding and dropped at scope end.
-        if let Some(dst) = self.try_lower_concat_str(value) {
-            self.value_of.insert(var, dst);
-            self.live_heap_handles.push(dst);
-            return Ok(true);
-        }
-        // `let ys = xs + [7]` — a SCALAR-element list concat EXECUTES to a fresh owned list
-        // (via the self-host __list_concat), held by the binding and dropped at scope end.
-        // The result is a REAL, POPULATED block (len(a)+len(b) copied slots), so a later
-        // `ys[i]` may index it directly. (A heap-element list concat returns None and falls
-        // through to the deferred Opaque.)
-        if let Some(dst) = self.try_lower_concat_list(value) {
-            self.value_of.insert(var, dst);
-            self.live_heap_handles.push(dst);
-            self.materialized_lists.insert(dst);
-            return Ok(true);
-        }
-        // `let s = "x=${n} y=${t}"` — a STRING INTERPOLATION over the executable
-        // subset (Lit / String Var/LitStr / Int Var/LitInt parts) EXECUTES to a
-        // fresh owned String via the same __str_concat chain, byte-matching v0;
-        // held by the binding and dropped at scope end. An interp with a compound
-        // (`${list}`) or call (`${f()}`) operand falls through to the deferred
-        // `Alloc{Opaque}` below, unchanged.
-        if let IrExprKind::StringInterp { parts } = &value.kind {
-            if let Some(dst) = self.try_lower_string_interp(parts) {
-                self.value_of.insert(var, dst);
-                self.live_heap_handles.push(dst);
-                return Ok(true);
-            }
-            // STRICT value mode: an interp the executable subset declined (a
-            // BLOCK-bodied operand — `${int.to_string({ let x = …; x * 3 })}` —
-            // or another non-lowerable piece) must NOT defer to the Opaque
-            // below: the binding reads back as an EMPTY string while native
-            // prints the real text — a silent wrong value on the verified
-            // default (the C-136 elide family, interp edition). REFUSE — the
-            // fn walls and v0 emits the correct bytes.
-            if crate::lower::strict_values() {
-                return Err(LowerError::Unsupported(
-                    "string interpolation outside the executable subset — \
-                     deferring it would read back an empty string not in this \
-                     brick"
-                        .into(),
-                ));
-            }
-        }
-        // `let xs = ["a" + "b", "c"]` — a List[String] literal with fresh-owned elements
-        // (the heap-container-element concat position; the −214 caps recovery).
-        if let Some(dst) = self.try_lower_str_list_literal(value) {
-            self.value_of.insert(var, dst);
-            self.live_heap_handles.push(dst);
-            return Ok(true);
-        }
-        // An Option ctor in the executable subset (`Some(scalar)` / `None`) is
-        // MATERIALIZED + tracked so a later `match` over the bound var executes;
-        // everything else is the deferred fresh `Alloc` (value-semantics).
-        if let Some(dst) = self.try_lower_option_ctor(value, ty) {
-            self.value_of.insert(var, dst);
-            self.live_heap_handles.push(dst);
-            return Ok(true);
-        }
-        Ok(false)
-    }
-
-    /// Extracted from `Self::lower_bind_heap_fresh` (second-round split, cog reduction):
-    /// the OptionSome/ResultOk/ResultErr honest-wall safety net, verbatim (the outer `if
-    /// let` guard now doubles as the "not applicable" fallthrough). `Ok(true)` means the
-    /// caller already bound `var` and should return immediately.
-    fn try_lower_bind_heap_fresh_variant_honest_wall(
-        &mut self,
-        var: VarId,
-        ty: &Ty,
-        value: &IrExpr,
-    ) -> Result<bool, LowerError> {
-        // HONEST-WALL SAFETY NET: a `some(<list>)` / `ok(<list>)` whose LIST payload the ctor
-        // materializer DECLINED (an exotic element the scalar/String/literal arms don't cover —
-        // e.g. a computed List[record]/List[List]) must NOT fall to the deferred Opaque `Alloc`
-        // below, which would read `none` / `ok([])` (the some(computed)/ok(computed) silent
-        // miscompile the adversarial fuzz surfaced). Wall instead — a wall is always safe, a
-        // wrong byte never is.
-        if let IrExprKind::OptionSome { expr }
-        | IrExprKind::ResultOk { expr }
-        | IrExprKind::ResultErr { expr } = &value.kind
-        {
-            use almide_lang::types::constructor::TypeConstructorId;
-            if matches!(&expr.ty,
-                Ty::Applied(TypeConstructorId::List, _) | Ty::Applied(TypeConstructorId::Map, _))
-            {
-                return Err(LowerError::Unsupported(
-                    "some/ok of a list or map payload outside the executable subset cannot be \
-                     faithfully materialized in this brick (e.g. an empty `[:]` — would defer \
-                     to an empty container)"
-                        .into(),
-                ));
-            }
-            // A CALL payload (`ok(result.unwrap_or(…))` — the C-149
-            // nested-share chain): ANF-materialize the call into a synth temp via
-            // the SAME `lower_bind` path a `let tmp = call` takes (tracked, typed
-            // drop, read shapes seeded — the lower_heap_extraction Call-container
-            // discipline), then rebuild the ctor over the temp VAR and retry the
-            // ctor materializer — its Var arms admit the payload with the share
-            // (Dup) discipline. A payload the Var arms still decline WALLS (the
-            // deferred Opaque would read `ok(0)` while native printed the err —
-            // the C-138 family; a wall is always safe, a wrong byte never).
-            // SCALAR call payloads take the same route (C-158): requiring a heap
-            // payload here let `ok(<un-lowerable Float combinator chain>)` skip
-            // to the deferred Opaque — an EMPTY block the formatter read as
-            // `ok(0)` while native printed the real value (differential-fuzz
-            // seed 1784512190387680000 index 74, a silent wrong value).
-            // TUPLE literal payloads too (`ok((0.3, 1000000))` — the (Float, Int)
-            // mixed-scalar payload the ctor materializer declines): deferred, the
-            // `result.flat_unwrap_or` twin read the EMPTY block's payload out of
-            // bounds — a runtime abort native never had (RunFailureDivergence).
-            // And IF-merged payloads (`ok(if c then some("a") else some("b"))`):
-            // deferred, the downstream unwrap misread the empty block as None and
-            // took the fallback while native returned the real payload — the ANF
-            // bind routes them through the proven heap-result-if bind machinery
-            // (C-106) instead.
-            if matches!(
-                expr.kind,
-                IrExprKind::Call { .. } | IrExprKind::Tuple { .. } | IrExprKind::If { .. }
-            ) {
-                let payload_ty = expr.ty.clone();
-                let payload = (**expr).clone();
-                let tmp = self.fresh_synth_var();
-                self.lower_bind(tmp, &payload_ty, &payload)?;
-                // The ANF bind itself may have DEFERRED (an Opaque skeleton —
-                // e.g. a capturing-closure element): retrying the ctor over it
-                // would wrap the same empty block. Wall instead.
-                if self
-                    .value_of
-                    .get(&tmp)
-                    .is_some_and(|v| self.deferred_opaque_binds.contains(v))
-                {
-                    return Err(LowerError::Unsupported(
-                        "some/ok/err payload whose materialization deferred cannot \
-                         be faithfully wrapped in this brick (walled, not read as a \
-                         zeroed ctor)"
-                            .into(),
-                    ));
-                }
-                let synth = IrExpr {
-                    kind: IrExprKind::Var { id: tmp },
-                    ty: payload_ty,
-                    span: payload.span,
-                    def_id: None,
-                };
-                let rebuilt_kind = match &value.kind {
-                    IrExprKind::OptionSome { .. } => {
-                        IrExprKind::OptionSome { expr: Box::new(synth) }
-                    }
-                    IrExprKind::ResultOk { .. } => {
-                        IrExprKind::ResultOk { expr: Box::new(synth) }
-                    }
-                    _ => IrExprKind::ResultErr { expr: Box::new(synth) },
-                };
-                let rebuilt = IrExpr {
-                    kind: rebuilt_kind,
-                    ty: value.ty.clone(),
-                    span: value.span,
-                    def_id: value.def_id,
-                };
-                if let Some(dst) = self.try_lower_option_ctor(&rebuilt, ty) {
-                    self.value_of.insert(var, dst);
-                    self.live_heap_handles.push(dst);
-                    return Ok(true);
-                }
-                return Err(LowerError::Unsupported(
-                    "some/ok/err of an un-admitted heap call payload cannot be \
-                     faithfully materialized in this brick (walled, not read as a \
-                     zeroed ctor)"
-                        .into(),
-                ));
-            }
-        }
-        Ok(false)
-    }
-
-    /// Extracted from `Self::lower_bind_heap_fresh` (second-round split, cog reduction):
-    /// the scalar/heap tuple construction, verbatim. `Ok(true)` means the caller already
-    /// bound `var` and should return immediately.
-    fn try_lower_bind_heap_fresh_tuple(&mut self, var: VarId, value: &IrExpr) -> Result<bool, LowerError> {
-        // A scalar-field tuple `(a, b)` of NON-LITERAL fields (vars / scalar exprs) — a
-        // literal `(3, 7)` is already an `Init::IntList` below. Construct the 2-slot block
-        // and store each field's computed value (the tuple-machinery construction sibling
-        // of the precise destructure). A heap-field tuple falls through to the Opaque alloc.
-        if let IrExprKind::Tuple { elements } = &value.kind {
-            if let Some(dst) = self.try_lower_scalar_tuple_construct(elements) {
-                self.value_of.insert(var, dst);
-                self.live_heap_handles.push(dst);
-                return Ok(true);
-            }
-            // A HEAP-element tuple (`(1, "a")`, `(p, 9)`) — materialize the mixed block
-            // + track its heap-slot mask, so `t.0`/`${tuple}` execute and the block (with
-            // its owned heap elements) is reclaimed by a masked recursive drop. Rolls back
-            // on a non-lowerable element (then Opaque → the Display walls).
-            let mark = self.ops.len();
-            let lhh_mark = self.live_heap_handles.len();
-            if let Some(dst) = self.try_lower_tuple_construct(elements) {
-                self.value_of.insert(var, dst);
-                self.live_heap_handles.push(dst);
-                return Ok(true);
-            }
-            self.ops.truncate(mark);
-            self.live_heap_handles.truncate(lhh_mark);
-        }
-        Ok(false)
-    }
-
-    /// Extracted from `Self::lower_bind_heap_fresh` (second-round split, cog reduction):
-    /// the scalar/heap record construction, verbatim. `Ok(true)` means the caller already
-    /// bound `var` and should return immediately.
-    fn try_lower_bind_heap_fresh_record(&mut self, var: VarId, value: &IrExpr) -> Result<bool, LowerError> {
-        // A SCALAR-only record `R { x: 3, y: 4 }` — build the tight-packed,
-        // width-aware block + store each field at its layout slot (the VALUE
-        // MODEL: `r.x`/`r.y` read back exactly what was stored). A HEAP-field
-        // record (a String/List field) needs an ownership-aware recursive drop
-        // this brick does not have, so it falls through to the deferred Opaque
-        // (which the field-access path then WALLS rather than mis-reads).
-        if let IrExprKind::Record { .. } = &value.kind {
-            if let Some(dst) = self.try_lower_scalar_record_construct(value) {
-                self.value_of.insert(var, dst);
-                self.live_heap_handles.push(dst);
-                return Ok(true);
-            }
-            // A VARIANT record-ctor literal (`Data { … }`) outside the builder's
-            // subset must WALL, not defer: a deferred Opaque variant read through a
-            // CALL arg (`tree_sum(t)`) bypasses the match-side deferred gate and the
-            // callee reads a garbage tag — the same miscompile class the Call-ctor
-            // bind gate above already errors on.
-            if let IrExprKind::Record { name: Some(n), .. } = &value.kind {
-                if self.variant_layouts.ctor_to_type.contains_key(n.as_str()) {
-                    return Err(LowerError::Unsupported(format!(
-                        "variant record-ctor `{}` bound to a let/var cannot be \
-                         faithfully materialized in this brick (a field outside the \
-                         ctor subset)",
-                        n.as_str()
-                    )));
-                }
-            }
-            // A record with one or more HEAP fields (`R { name: "x", n: i }`) —
-            // materialize the mixed scalar+heap block + track its heap-slot mask, so a
-            // `r.n` scalar read AND a `r.name` heap read execute and the block (with its
-            // owned heap fields) is reclaimed by a masked recursive drop. Rolls back on
-            // a partially-lowered out-of-subset field (a heap-returning-call field).
-            let mark = self.ops.len();
-            let lhh_mark = self.live_heap_handles.len();
-            if let Some(dst) = self.try_lower_record_construct(value) {
-                self.value_of.insert(var, dst);
-                self.live_heap_handles.push(dst);
-                return Ok(true);
-            }
-            self.ops.truncate(mark);
-            self.live_heap_handles.truncate(lhh_mark);
-        }
-        Ok(false)
-    }
-
-    /// Extracted from `Self::lower_bind_heap_fresh` (second-round split, cog reduction):
-    /// the spread-record construction, verbatim. `Ok(true)` means the caller already
-    /// bound `var` and should return immediately.
-    fn try_lower_bind_heap_fresh_spread_record(
-        &mut self,
-        var: VarId,
-        value: &IrExpr,
-    ) -> Result<bool, LowerError> {
-        // A SPREAD record `R { ...base, f: override }` — build a fresh block of the
-        // same layout, COPYING each non-overridden field from `base` (a scalar load,
-        // a heap-handle Dup so both records own a distinct reference) and storing the
-        // overrides. So `let b2 = Box { ...b, value: 8 }` reads `b2.value=8
-        // b2.label=old` while `b.label` still reads `old`. Rolls back to the deferred
-        // Opaque (whose field reads WALL) on a non-materialized base / out-of-subset
-        // override — never wrong bytes.
-        if let IrExprKind::SpreadRecord { .. } = &value.kind {
-            let mark = self.ops.len();
-            let lhh_mark = self.live_heap_handles.len();
-            if let Some(dst) = self.try_lower_spread_record_construct(value) {
-                self.value_of.insert(var, dst);
-                self.live_heap_handles.push(dst);
-                return Ok(true);
-            }
-            self.ops.truncate(mark);
-            self.live_heap_handles.truncate(lhh_mark);
-        }
-        Ok(false)
-    }
-
-    /// Extracted from `Self::lower_bind_heap_fresh` (second-round split, cog reduction):
-    /// the scalar-list construct + non-empty heap-element-list honest wall, verbatim.
-    /// `Ok(true)` means the caller already bound `var` and should return immediately.
-    fn try_lower_bind_heap_fresh_scalar_list(
-        &mut self,
-        var: VarId,
-        ty: &Ty,
-        value: &IrExpr,
-    ) -> Result<bool, LowerError> {
-        // A scalar `List[Int/Float/Bool]` literal with COMPUTED elements (`[1.0, inf, 0.5]`,
-        // `[a, a]`) — build the block + store each slot (an all-literal list is the IntList
-        // path in `alloc_init` below; a computed element can't fold to a constant).
-        if let Some(dst) = self.try_lower_scalar_list_construct(value) {
-            self.value_of.insert(var, dst);
-            self.live_heap_handles.push(dst);
-            return Ok(true);
-        }
-        // A NON-EMPTY `List[heap]` LITERAL that NONE of the materialization paths above
-        // could build — a list of heap-FIELD records/tuples (`[R{name:String,…}, …]`), a
-        // list of lists, a list of heap call-results. The flat `Init::Opaque` fallback
-        // below would emit an EMPTY len-0 block (`list_new(0, …)`); a later `list.map` /
-        // `list.sort_by` / `xs[i]` over it then silently reads NOTHING = wrong/empty bytes.
-        // (A heap-field-record element needs a TWO-LEVEL recursive drop — the list frees
-        // each record, each record frees its String fields — which the single-level
-        // `DropListStr` cannot express without a new ownership op; that is the
-        // nested-ownership frontier, out of this brick.) WALL the function cleanly instead
-        // of mis-valuing it — the render discards it (no invalid wasm, no empty output).
-        // GATED to a NON-EMPTY heap-element `List` LITERAL (an empty `[]`, a scalar list,
-        // and a `List[String]`/scalar-aggregate list are all handled above), so this only
-        // rejects the genuinely-unmaterializable case.
-        if let IrExprKind::List { elements } = &value.kind {
-            use almide_lang::types::constructor::TypeConstructorId;
-            let heap_elem_list = matches!(ty,
-                Ty::Applied(TypeConstructorId::List, a) if a.len() == 1 && is_heap_ty(&a[0]));
-            if heap_elem_list && !elements.is_empty() {
-                // A List[Record] literal materializes via the record-list builder (drop →
-                // $__drop_list_<R>); other nested-ownership element lists stay walled.
-                if let Some(dst) = self.try_lower_record_list_literal(value) {
-                    self.value_of.insert(var, dst);
-                    return Ok(true);
-                }
-                return Err(LowerError::Unsupported(
-                    "non-empty List[heap] literal with nested-ownership elements \
-                     (a heap-field record/tuple, a list, a call result) cannot be \
-                     faithfully materialized in this brick (walled, not emitted empty)"
-                        .into(),
-                ));
-            }
-        }
-        Ok(false)
-    }
-
-    /// Extracted from `Self::lower_bind_heap_fresh` (second-round split, cog reduction):
-    /// the terminal `Alloc{Opaque}`-or-real-list fallback, verbatim.
-    fn lower_bind_heap_fresh_opaque(&mut self, var: VarId, ty: &Ty, value: &IrExpr) -> Result<(), LowerError> {
-        let dst = self.fresh_value();
-        let repr = repr_of(ty)?;
-        let init = alloc_init(value);
-        // A NON-EMPTY SCALAR-element List literal that did NOT fold to a real
-        // `Init::IntList` (a computed element the builders above declined — e.g. a
-        // nested inadmissible-HOF chain, fuzz B-198/659): the flat Opaque reads as
-        // an EMPTY list at any observation (`${r4}` printed `[]` while native
-        // printed the values — the silent-wrong-value class). WALL instead — the
-        // heap-element twin at the gate above already does; this is its scalar twin.
-        if let IrExprKind::List { elements } = &value.kind {
-            use almide_lang::types::constructor::TypeConstructorId;
-            let scalar_elem_list = matches!(ty,
-                Ty::Applied(TypeConstructorId::List, a) if a.len() == 1 && !is_heap_ty(&a[0]));
-            if scalar_elem_list && !elements.is_empty() && matches!(init, Init::Opaque) {
-                return Err(LowerError::Unsupported(
-                    "non-empty scalar-element list literal with an element outside \
-                     the executable subset cannot be faithfully materialized in this \
-                     brick (walled, not emitted empty)"
-                        .into(),
-                ));
-            }
-        }
-        // An all-literal `Init::IntList` is a REAL, POPULATED block (every slot a constant) —
-        // admit a direct `xs[i]` bounds-checked load over it. An `Init::Opaque` (a deferred /
-        // unsupported value) is NOT tracked: indexing it would trap on cap 0.
-        let real_list = matches!(init, Init::IntList(_));
-        self.value_of.insert(var, dst);
-        self.ops.push(Op::Alloc { dst, repr, init });
-        self.live_heap_handles.push(dst);
-        if real_list {
-            self.materialized_lists.insert(dst);
-        }
-        self.record_elided_calls(value);
-        Ok(())
-    }
-
-    /// Extracted from `Self::lower_bind_heap` (pattern-2 uniform-arm split, cog reduction):
-    /// the arm body verbatim, re-narrowed via `let-else`. Pure text move.
-    fn lower_bind_heap_call_named(&mut self, var: VarId, ty: &Ty, value: &IrExpr) -> Result<(), LowerError> {
-        let IrExprKind::Call { target: CallTarget::Named { name }, args, .. } = &value.kind else { unreachable!() };
-        // A custom-variant CONSTRUCTOR `let t = Num(9)` (ADT brick 2) is NOT a call —
-        // build the tagged value-model block (tag@slot0 + scalar fields@slot1..), bound
-        // + dropped at scope end (cert `i` + `d`, like the scalar-record bind). Must
-        // precede the CallFn emission, which would emit a dangling `(call $Num)`. A
-        // heap/recursive ctor field is ADT brick 5 → WALL (never a wrong-bytes block).
-        if self.variant_layouts.ctor_to_type.contains_key(name.as_str()) {
-            if let Some(dst) = self.try_lower_variant_ctor(value) {
-                self.value_of.insert(var, dst);
-                self.live_heap_handles.push(dst);
-                return Ok(());
-            }
-            return Err(LowerError::Unsupported(format!(
-                "variant constructor `{}` bound to a let/var cannot be faithfully \
-                 materialized in this brick (a heap/recursive field — ADT brick 5)",
-                name.as_str()
-            )));
-        }
-        let lowered = self.lower_call_args(args)?;
-        let dst = self.fresh_value();
-        // A function-VALUED result (`let f = mk()`) is a CLOSURE BLOCK — the uniform
-        // heap representation (`repr_of(Ty::Fn)` = Ptr), owned + dropped at scope end
-        // like any heap result; `closure_values` (below) makes a later `f(args)`
-        // dispatch through it.
-        let repr = repr_of(ty)?;
-        self.value_of.insert(var, dst);
-        self.ops.push(Op::CallFn {
-            dst: Some(dst),
-            name: name.as_str().to_string(),
-            args: lowered,
-            result: Some(repr),
-        });
-        self.live_heap_handles.push(dst);
-        self.seed_call_named_heap_drop_route(dst, ty);
-        self.seed_call_named_heap_read_shape(dst, ty);
-        Ok(())
-    }
-
-    /// Extracted from `Self::lower_bind_heap_call_named` (second-round split, cog
-    /// reduction): the mutually-exclusive drop-route selection for a Named-call's fresh
-    /// heap result, verbatim (the original `if/else if` chain).
-    fn seed_call_named_heap_drop_route(&mut self, dst: ValueId, ty: &Ty) {
-        if !self.seed_call_named_heap_drop_route_a(dst, ty) {
-            self.seed_call_named_heap_drop_route_b(dst, ty);
-        }
-    }
-
-    /// Extracted from `Self::seed_call_named_heap_drop_route` (third-round split, cog
-    /// reduction): the first half of the mutually-exclusive `if/else if` chain, verbatim.
-    /// Returns whether a branch matched (the caller then skips the second half).
-    fn seed_call_named_heap_drop_route_a(&mut self, dst: ValueId, ty: &Ty) -> bool {
-        if crate::lower::is_res_intlist_strlist_ty(ty) {
-            // `result.collect` — Result[List[Int], List[String]]: the TAG-AWARE
-            // generated `$__drop_res_ilsl` (Err → recursive string free, Ok → flat;
-            // either flat class would leak or double-free one side).
-            self.variant_drop_handles.insert(dst, "res_ilsl".to_string());
-            self.materialized_results_str.insert(dst);
-        } else if crate::lower::is_list_list_str_ty(ty) {
-            self.list_list_str_lists.insert(dst);
-        } else if crate::lower::is_list_str_str_ty(ty) {
-            // `List[(String,String)]` (map.entries) — DropListStrStr frees each tuple's two
-            // Strings; the flat heap_elem_lists DropListStr would leak them (a render loop OOMs).
-            self.str_str_elem_lists.insert(dst);
-        } else if crate::lower::is_list_int_str_ty(ty) {
-            // `List[(Int,String)]` (list.enumerate) — recursive `$__drop_list_int_str` (rc_dec
-            // each tuple's String); the flat heap_elem_lists DropListStr would leak them.
-            self.variant_drop_handles.insert(dst, "list_int_str".to_string());
-        } else if crate::lower::is_map_ivh_ty(ty) {
-            // `Map[Int, String]` — `$__drop_map_ivh` rc_decs each OWNED value slot.
-            self.variant_drop_handles.insert(dst, "map_ivh".to_string());
-        } else if crate::lower::is_map_fn_ty(ty) {
-            // `Map[String, <Fn>]` — `$__drop_map_mclo` frees each value via
-            // `__drop_closure` (the hval flat rc_dec would leak captured env).
-            self.variant_drop_handles.insert(dst, "map_mclo".to_string());
-        } else if crate::lower::is_map_hval_ty(ty) {
-            // `Map[String, List[scalar]]` — `$__drop_map_hval` rc_decs all 2n slots.
-            self.variant_drop_handles.insert(dst, "map_hval".to_string());
-        } else if let Some(hname) = self.map_named_value_drop(ty) {
-            // `Map[String, <record/variant>]` — the desugared map literal's
-            // from_list result (type-driven sweep; see `map_named_value_drop`).
-            self.variant_drop_handles.insert(dst, hname);
-        } else if crate::lower::is_map_msv_ty(ty) {
-            // `Map[String, Map[String, String]]` — `$__drop_map_msv` sweeps each
-            // last-ref inner map's String slots (a flat rc_dec would leak them).
-            self.variant_drop_handles.insert(dst, "map_msv".to_string());
-        } else {
-            return false;
-        }
-        true
-    }
-
-    /// Extracted from `Self::seed_call_named_heap_drop_route` (third-round split, cog
-    /// reduction): the second half of the mutually-exclusive `if/else if` chain, verbatim
-    /// (only reached when the first half's chain did not match).
-    fn seed_call_named_heap_drop_route_b(&mut self, dst: ValueId, ty: &Ty) {
-        if crate::lower::is_map_mlo_ty(ty) {
-            // `Map[String, List[Option[Int]]]` — `$__drop_map_mlo` sweeps each
-            // last-ref value list's Option slots (a flat rc_dec would leak them).
-            self.variant_drop_handles.insert(dst, "map_mlo".to_string());
-        } else if let Some(rname) = (match ty {
-            Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, a)
-                if a.len() == 1 =>
-            {
-                self.record_or_anon_drop_type_name(&a[0])
-            }
-            _ => None,
-        }) {
-            // A `List[<recursive-drop record>]` result (`list.unique` over a
-            // String-field record via the `__krec_*` twins): route to the generated
-            // `$__drop_list_<R>` (emitted for EVERY recursive-drop record) — the
-            // flat per-slot dec freed each element block but LEAKED its String
-            // fields (the krec-unique residue).
-            self.variant_drop_handles.insert(dst, format!("list_{rname}"));
-        } else if crate::lower::is_lenlist_list_ty(ty) {
-            // `List[Result[_, String]]`/`List[Option[String]]` — the len-loop drop; the
-            // flat DropListStr would leak each element's owned payload slots.
-            self.variant_drop_handles.insert(dst, "list_lenlist".to_string());
-        } else if crate::lower::is_opt_list_str_ty(ty) {
-            // `Option[List[String]]` (the heap-acc fold value) — physically a 0/1-element
-            // List[List[String]]; the nested DropListListStr sweep is its exact free (the
-            // flat DropListStr would leak the stack Strings).
-            self.list_list_str_lists.insert(dst);
-        } else if matches!(ty,
-            Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Map, a)
-                if a.len() == 2 && matches!(a[0], Ty::String) && !is_heap_ty(&a[1]))
-        {
-            // `Map[String, <scalar>]` (split layout, @4 = n): the DropListStr sweep
-            // rc_decs exactly the n deep-copied key Strings (scalar value slots
-            // untouched) — the bare flat rc_dec LEAKED every key copy per bind (a
-            // latent leak the map.fold heap-acc loop made observable at a 4MB cap).
-            self.heap_elem_lists.insert(dst);
-        } else if is_heap_elem_list_ty(ty) {
-            self.heap_elem_lists.insert(dst);
-        } else if is_scalar_elem_list_ty(ty) {
-            // A user fn returning `List[scalar]` yields a REAL, POPULATED list
-            // block (the v1 calling convention — the same argument as the
-            // variant/record seeds below; a callee that cannot build one WALLS,
-            // and the render rejects the program). Admit a direct `xs[i]`
-            // bounds-checked load over the bound result — the C-132 move-mode
-            // write-back binds the returned buffer exactly here (`__mp_buf =
-            // add_item(data, 1)` then `data = __mp_buf; data[0]`).
-            self.materialized_lists.insert(dst);
-        }
-    }
-
-    /// Extracted from `Self::lower_bind_heap_call_named` (second-round split, cog
-    /// reduction): the independent (non-mutually-exclusive) read-shape tracking checks
-    /// for a Named-call's fresh heap result, verbatim.
-    fn seed_call_named_heap_read_shape(&mut self, dst: ValueId, ty: &Ty) {
-        // A user fn returning `List[heap]` (`build_nested() -> List[List[Int]]`) is
-        // likewise a REAL, POPULATED nested-ownership block, so admit the element
-        // borrow `nested[i]` (LoadHandle at `$elem_addr`) over the bound var — the
-        // exact mirror of the Module-call bind's `string.split` registration. Without
-        // it, `nested[10]` fell to the container-grain Dup of the WHOLE list and a
-        // consumer read the outer block (`list.len(mid)` = the OUTER len — the
-        // rc_alloc_stress silent miscompile). Narrowed to `List[heap]` (NOT the
-        // broader Option/Result/Map `is_heap_elem_list_ty` also matches) — only a
-        // real list is `[i]`-indexable here.
-        if matches!(ty, Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, a)
-                if a.len() == 1 && is_heap_ty(&a[0]))
-        {
-            self.materialized_lists.insert(dst);
-        }
-        // A `Value` result from a user fn (`let v = parse_number(c, raw)`) drops via the
-        // runtime-tag-dispatched `DropValue` — the SAME marking the Module-call bind path does
-        // (was missing here, so a let-bound Named-call Value leaked: a parse loop OOMs).
-        if crate::lower::is_value_ty(ty) {
-            self.value_handles.insert(dst);
-        }
-        // A user fn RETURNING a function value (`let f = mk()` / `let f = adder(3)`)
-        // yields a CLOSURE BLOCK — a fresh owned heap value (already in the scope-end
-        // set like any heap result): track it so a later `f(args)` dispatches through
-        // `Op::CallIndirect` via `emit_closure_call`.
-        if matches!(ty, Ty::Fn { .. }) {
-            self.closure_values.insert(dst);
-        }
-        // A user function returning Option/Result yields a REAL same-layout variant block
-        // (the v1 calling convention — `seed_variant_param`'s contract). SEED its READ-shape
-        // so a later `match x { … }` / `x ?? d` over the LET-BOUND var EXECUTES (reads the
-        // tag) exactly as the direct-call-arg position already does (`lower_call_args`'s
-        // Named arm). Adds ONLY layout knowledge — `dst` is already an owned heap value
-        // dropped at scope end, so no ownership/cert change. This is what made
-        // `let parsed = parse_oct(d); match parsed { … }` (num_signed_base, after the
-        // let-bound-heap-`if` tail-duplication) lower instead of wall.
-        if is_variant_ty(ty) {
-            self.seed_variant_param(dst, ty);
-        } else if let Some((_, tys)) = self.aggregate_field_tys(ty) {
-            // A user function returning a RECORD/TUPLE yields a REAL same-layout block (the
-            // callee built it via try_lower_record_construct). Seed its READ-shape
-            // (materialized_aggregates) so a field read `p.y` loads the real slot instead of
-            // falling back to the container-grain Dup (which returns the whole record — the
-            // `mk(5).y` empty-string miscompile), AND its heap-slot MASK (record_masks) so the
-            // OWNED scope-end drop frees exactly the heap fields (no leak, no double-free).
-            let heap_slots: Vec<usize> =
-                (0..tys.len()).filter(|&i| is_heap_ty(&tys[i])).collect();
-            self.materialized_aggregates.insert(dst);
-            self.record_masks.insert(dst, heap_slots);
-            // A record with a Map/List[heap]/record/Value field drops RECURSIVELY ($__drop_<R>),
-            // not the flat masked DropListStr (which would leak the nested heap) — route it. An
-            // ANONYMOUS record return whose flat one-level mask would leak a nested heap field
-            // (`{ data: Bytes, state: Cfb8State }` — aes cfb8) routes to its synthesized
-            // `__drop_anonrec_<hash>` (so `state` is freed through `__drop_Cfb8State`).
-            if let Some(name) = self.record_or_anon_drop_type_name(ty) {
-                self.variant_drop_handles.insert(dst, name);
-            }
-        }
-    }
-
-    /// Extracted from `Self::lower_bind_heap` (pattern-2 uniform-arm split, cog reduction):
-    /// the arm body verbatim, re-narrowed via `let-else`. Pure text move.
-    fn lower_bind_heap_call_module(&mut self, var: VarId, ty: &Ty, value: &IrExpr) -> Result<(), LowerError> {
-        let IrExprKind::Call { target: CallTarget::Module { module, func, .. }, args, .. } = &value.kind else { unreachable!() };
-        let dst =
-            self.lower_pure_module_value_call(module.as_str(), func.as_str(), args, ty)?;
-        self.value_of.insert(var, dst);
-        let faithful = self.check_call_module_faithful(module.as_str(), func.as_str(), args)?;
-        self.seed_call_module_heap_read_shape(dst, ty, module.as_str(), func.as_str(), faithful);
-        self.seed_call_module_heap_drop_route(dst, ty);
-        // A `Value` result (value.str/int/… or a Value-returning combinator) drops via the
-        // runtime-tag-dispatched DropValue (a heap-payload Value owns one handle).
-        if crate::lower::is_value_ty(ty) {
-            self.value_handles.insert(dst);
-        }
-        Ok(())
-    }
-
-    /// Extracted from `Self::lower_bind_heap_call_module` (second-round split, cog
-    /// reduction): the HOF-faithfulness guard, verbatim. Returns whether the call is
-    /// FAITHFULLY executable (every closure arg lifted, no un-representable fn-typed data
-    /// arg); an unfaithful higher-order call still WALLS via `Err`, exactly as before.
-    fn check_call_module_faithful(
-        &mut self,
-        module: &str,
-        func: &str,
-        args: &[IrExpr],
-    ) -> Result<bool, LowerError> {
-        // A SCALAR-element `List[Int/Float/Bool]` result from a self-host list call is a REAL,
-        // POPULATED block — admit a direct `xs[i]` — ONLY when the call is FAITHFULLY executable:
-        //  (1) every closure arg LIFTED (an unlifted `list.map(fns, (f) => f(10))` runs the
-        //      combinator with a missing slot → empty/garbage), AND
-        //  (2) no DATA argument carries an UN-REPRESENTABLE function type (this comment
-        //      historically said "list.map(fns, …) over fns: List[(Int)->Int] — a list of
-        //      closures the v1 model cannot represent" — no longer true: B36 shipped
-        //      `List[<Fn>]` literal construction + a generated per-element `$__drop_
-        //      list_closure`, so a `List[Fn]` DATA arg is now a REAL, populated,
-        //      correctly-freed block — excluded below). A Fn buried in some OTHER shape
-        //      (a record/tuple field, a nested nested-List[List[Fn]]) is still unrepresented
-        //      and stays walled. The combinator's OWN closure arg (a `Lambda`/`FnRef`,
-        //      function-typed by construction) is EXCLUDED too — it is handled by (1), and
-        //      `(p) => p.x` over `points: List[Point]` is the faithful case that must stay
-        //      tracked.
-        // Otherwise the result is unmaterialized and a `xs[i]` over it would TRAP on cap 0, so
-        // it is left deferring to `Const 0` (mis-valued, never a new runtime crash).
-        let data_arg_has_fn = args.iter().any(|a| {
-            // A let-bound lambda passed BY NAME (`let g = (x) => …; xs |> list.map(g)`) is a
-            // CLOSURE arg — try_lower_defunc_list_hof resolves it via lambda_bindings and inlines
-            // it faithfully (calls.rs). Without recognizing the `Var` here it is misread as a
-            // fn-typed DATA arg (a `list.map(fns, …)` over a list-of-closures the v1 model can't
-            // represent) and the guard below WALLS it — even though the inline succeeded. This is
-            // the bind-vs-tail discrepancy: the tail/value position has no such data-arg guard, so
-            // `let g = …; xs |> map(g)` lowered as a TAIL but walled as a `let r = xs |> map(g)`.
-            let is_closure_arg = matches!(
-                &a.kind,
-                IrExprKind::Lambda { .. } | IrExprKind::FnRef { .. } | IrExprKind::ClosureCreate { .. }
-            ) || matches!(&a.kind, IrExprKind::Var { id } if self.lambda_bindings.contains_key(id));
-            // `List[<Fn>]` (B36) and `Map[String, <Fn>]` (the mclo family — the
-            // hval handle-level twins + `$__drop_map_mclo`) are representable
-            // closure CONTAINERS — excluded from the wall.
-            let is_representable_closure_list = matches!(&a.ty,
-                Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, e)
-                    if e.len() == 1 && matches!(e[0], Ty::Fn { .. }))
-                || crate::lower::is_map_fn_ty(&a.ty);
-            !is_closure_arg && !is_representable_closure_list && crate::lower::ty_contains_fn(&a.ty)
-        });
-        let faithful = !self.last_call_had_unlifted_closure && !data_arg_has_fn;
-        // WALL the UNFAITHFUL higher-order combinator instead of silently
-        // mis-valuing it. A HOF call (`list.map`/`filter`/`fold`…) over a
-        // CAPTURING/param-invoking lambda (no liftable slot) or a fn-typed DATA
-        // arg (`list.map(fns, (f) => f(10))` over `fns: List[(Int)->Int]` — a
-        // list of closures the v1 model cannot represent) runs the self-host
-        // combinator with a missing/garbage closure and produces a zero-filled
-        // result. Leaving the result deferred (a `Const 0` `xs[i]`) emits WRONG
-        // BYTES — a silent miscompile. Walling the whole function here is the
-        // honest outcome (render discards it cleanly; no invalid wasm, no wrong
-        // output). The FAITHFUL case (every closure lifted, no fn-typed data —
-        // `list.map(xs, (x) => x + 1)`, `(p) => p.x` over `List[Point]`) is
-        // UNTOUCHED, so the in-scope HOF byte-matches stay materialized.
-        if crate::lower::is_higher_order(args) && !faithful {
-            if std::env::var("ALMIDE_DBG_ANF").is_ok() {
-                eprintln!(
-                    "[hof-guard] {}.{} unlifted={} data_arg_has_fn={}",
-                    module,
-                    func,
-                    self.last_call_had_unlifted_closure,
-                    data_arg_has_fn
-                );
-            }
-            return Err(LowerError::Unsupported(format!(
-                "{module}.{func} with an unliftable/closure-list higher-order argument \
-                 cannot execute faithfully in this brick (walled, not mis-valued)"
-            )));
-        }
-        Ok(faithful)
-    }
-
-    /// Extracted from `Self::lower_bind_heap_call_module` (second-round split, cog
-    /// reduction): the `faithful`-gated + self-host fn-name read-shape tracking,
-    /// verbatim.
-    fn seed_call_module_heap_read_shape(
-        &mut self,
-        dst: ValueId,
-        ty: &Ty,
-        module: &str,
-        func: &str,
-        faithful: bool,
-    ) {
-        if is_scalar_elem_list_ty(ty) && faithful {
-            self.materialized_lists.insert(dst);
-        }
-        // A faithful `List[heap]` result (`string.split`/`chars`/`lines` → `List[String]`,
-        // or a heap-element list combinator) is ALSO a REAL, POPULATED nested-ownership block
-        // whose slots hold owned element HANDLES — so a value-position `xs[i]` over the bound
-        // var can LoadHandle element i at `$elem_addr` (the heap-element borrow path in
-        // `try_lower_heap_field_borrow`, gated on `materialized_lists`). Without registering
-        // it, `parts[i]` fell to the container-grain `Dup` of the WHOLE list → a String
-        // consumer read the list HEADER bytes (the `string.split`-subscript miscompiles).
-        // Narrowed to `List[heap]` (NOT the broader Option/Result/Map that
-        // `is_heap_elem_list_ty` also matches) — only a real list is `[i]`-indexable here.
-        if matches!(ty, Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, a)
-                if a.len() == 1 && is_heap_ty(&a[0]))
-            && faithful
-        {
-            self.materialized_lists.insert(dst);
-        }
-        // A self-host returning a RECORD/TUPLE (`list.partition` → (List, List)):
-        // seed the READ-shape — without it a `.0`/`.1` projection falls to the
-        // container-grain Dup and a consumer reads the TUPLE header
-        // (list.len(result.0) returned the tuple's len 2, not the slot list's 5 —
-        // the pipe_chain partition miscompile, 2026-07-17). READ-shape ONLY: the
-        // drop stays the pre-existing flat one (re-routing it through record_masks
-        // here imbalanced the ownership cert — the callee's fills are opaque to
-        // the caller's witness; the Named arm's mask rides a different accounting).
-        if faithful && self.aggregate_field_tys(ty).is_some() {
-            self.materialized_aggregates.insert(dst);
-        }
-        // A BORROW result (`prim.load_str` of a list slot — the list still owns it) is NOT
-        // added to the scope-end drop set; everything else is a fresh owned value.
-        if !self.param_values.contains(&dst) {
-            self.live_heap_handles.push(dst);
-        }
-        // A self-host Option fn (`list.get`) returns a real materialized Option —
-        // track the bound result so a later `match` over the var EXECUTES.
-        if is_self_host_option_module_fn(module, func) {
-            self.materialized_options.insert(dst);
-        }
-        // A FUNCTION-valued module-call result (`let f = map.get_or(m, k, d)` —
-        // the closure-valued map read): the result IS a closure block — track it
-        // so a later `f()` dispatches via CallIndirect, and its scope-end drop
-        // routes to the recursive `$__drop_closure` (`closure_values` drives
-        // `drop_op_for`; a captured env slot would leak under the flat rc_dec).
-        if matches!(ty, Ty::Fn { .. }) {
-            self.closure_values.insert(dst);
-        }
-        // A self-host Result fn (`int.parse`) returns a real materialized Result — track it
-        // so a later `match r { Ok(v) => …, Err(e) => … }` over the var EXECUTES.
-        if is_self_host_result_module_fn(module, func) {
-            self.materialized_results.insert(dst);
-        }
-        // A self-host HEAP-Ok Result fn (`value.as_string`/`value.as_array`) — track it in the
-        // cap-as-tag set so a `match` reads tag @16 + binds the @12 payload. The DROP differs
-        // by Ok-arm: a `List[Value]` Ok (`value.as_array`) frees recursively
-        // (`value_result_lists` → `DropResultListValue`), else a String Ok flat (`DropListStr`).
-        if crate::lower::is_self_host_result_str_module_fn(module, func) {
-            self.materialized_results_str.insert(dst);
-            if crate::lower::is_result_listval_ty(ty) {
-                self.value_result_lists.insert(dst);
-            } else if crate::lower::is_value_result_ty(ty) {
-                // `Result[Value, String]` (value.get) — a single dynamic Value Ok, freed
-                // recursively by `Op::DropResultValue` (Ok → `$__drop_value`).
-                self.value_result_results.insert(dst);
-            } else {
-                self.heap_elem_lists.insert(dst);
-            }
-        }
-    }
-
-    /// Extracted from `Self::lower_bind_heap_call_module` (second-round split, cog
-    /// reduction): the mutually-exclusive drop-route selection for a Module-call's fresh
-    /// heap result, verbatim (the original `if/else if` chain — NOT the same helper as
-    /// `Self::seed_call_named_heap_drop_route`: this one omits the `is_scalar_elem_list_ty`
-    /// tail arm, already handled above via the `faithful` gate).
-    fn seed_call_module_heap_drop_route(&mut self, dst: ValueId, ty: &Ty) {
-        if !self.seed_call_module_heap_drop_route_a(dst, ty) {
-            self.seed_call_module_heap_drop_route_b(dst, ty);
-        }
-    }
-
-    /// Extracted from `Self::seed_call_module_heap_drop_route` (third-round split, cog
-    /// reduction): the first half of the mutually-exclusive `if/else if` chain, verbatim.
-    /// Returns whether a branch matched (the caller then skips the second half).
-    fn seed_call_module_heap_drop_route_a(&mut self, dst: ValueId, ty: &Ty) -> bool {
-        // A `List[String]` result (string.split / a List[String] combinator) is a
-        // nested-ownership list — its scope-end drop must recursively free elements.
-        if crate::lower::is_res_intlist_strlist_ty(ty) {
-            // `result.collect` — Result[List[Int], List[String]]: the TAG-AWARE
-            // generated `$__drop_res_ilsl` (Err → recursive string free, Ok → flat;
-            // either flat class would leak or double-free one side).
-            self.variant_drop_handles.insert(dst, "res_ilsl".to_string());
-            self.materialized_results_str.insert(dst);
-        } else if crate::lower::is_list_list_str_ty(ty) {
-            self.list_list_str_lists.insert(dst);
-        } else if crate::lower::is_list_str_str_ty(ty) {
-            // `List[(String,String)]` (map.entries) — DropListStrStr frees each tuple's two
-            // Strings; the flat heap_elem_lists DropListStr would leak them (a render loop OOMs).
-            self.str_str_elem_lists.insert(dst);
-        } else if crate::lower::is_list_int_str_ty(ty) {
-            // `List[(Int,String)]` (list.enumerate) — recursive `$__drop_list_int_str`; the flat
-            // heap_elem_lists DropListStr would leak each tuple's String (a 10⁴ loop OOMs).
-            self.variant_drop_handles.insert(dst, "list_int_str".to_string());
-        } else if crate::lower::is_map_ivh_ty(ty) {
-            // `Map[Int, String]` — `$__drop_map_ivh` rc_decs each OWNED value slot.
-            self.variant_drop_handles.insert(dst, "map_ivh".to_string());
-        } else if crate::lower::is_map_fn_ty(ty) {
-            // `Map[String, <Fn>]` — `$__drop_map_mclo` frees each value via
-            // `__drop_closure` (the hval flat rc_dec would leak captured env).
-            self.variant_drop_handles.insert(dst, "map_mclo".to_string());
-        } else if crate::lower::is_map_hval_ty(ty) {
-            // `Map[String, List[scalar]]` — `$__drop_map_hval` rc_decs all 2n slots.
-            self.variant_drop_handles.insert(dst, "map_hval".to_string());
-        } else if let Some(hname) = self.map_named_value_drop(ty) {
-            // `Map[String, <record/variant>]` — the desugared map literal's
-            // from_list result (type-driven sweep; see `map_named_value_drop`).
-            self.variant_drop_handles.insert(dst, hname);
-        } else {
-            return false;
-        }
-        true
-    }
-
-    /// Extracted from `Self::seed_call_module_heap_drop_route` (third-round split, cog
-    /// reduction): the second half of the mutually-exclusive `if/else if` chain, verbatim
-    /// (only reached when the first half's chain did not match).
-    fn seed_call_module_heap_drop_route_b(&mut self, dst: ValueId, ty: &Ty) {
-        if crate::lower::is_map_msv_ty(ty) {
-            // `Map[String, Map[String, String]]` — `$__drop_map_msv` sweeps each
-            // last-ref inner map's String slots (a flat rc_dec would leak them).
-            self.variant_drop_handles.insert(dst, "map_msv".to_string());
-        } else if crate::lower::is_map_mlo_ty(ty) {
-            // `Map[String, List[Option[Int]]]` — `$__drop_map_mlo` sweeps each
-            // last-ref value list's Option slots (a flat rc_dec would leak them).
-            self.variant_drop_handles.insert(dst, "map_mlo".to_string());
-        } else if let Some(rname) = (match ty {
-            Ty::Applied(almide_lang::types::constructor::TypeConstructorId::List, a)
-                if a.len() == 1 =>
-            {
-                self.record_or_anon_drop_type_name(&a[0])
-            }
-            _ => None,
-        }) {
-            // A `List[<recursive-drop record>]` result (`list.unique` over a
-            // String-field record via the `__krec_*` twins): route to the generated
-            // `$__drop_list_<R>` (emitted for EVERY recursive-drop record) — the
-            // flat per-slot dec freed each element block but LEAKED its String
-            // fields (the krec-unique residue).
-            self.variant_drop_handles.insert(dst, format!("list_{rname}"));
-        } else if crate::lower::is_lenlist_list_ty(ty) {
-            // `List[Result[_, String]]`/`List[Option[String]]` — the len-loop drop; the
-            // flat DropListStr would leak each element's owned payload slots.
-            self.variant_drop_handles.insert(dst, "list_lenlist".to_string());
-        } else if crate::lower::is_opt_list_str_ty(ty) {
-            // `Option[List[String]]` (the heap-acc fold value) — physically a 0/1-element
-            // List[List[String]]; the nested DropListListStr sweep is its exact free (the
-            // flat DropListStr would leak the stack Strings).
-            self.list_list_str_lists.insert(dst);
-        } else if matches!(ty,
-            Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Map, a)
-                if a.len() == 2 && matches!(a[0], Ty::String) && !is_heap_ty(&a[1]))
-        {
-            // `Map[String, <scalar>]` (split layout, @4 = n): the DropListStr sweep
-            // rc_decs exactly the n deep-copied key Strings (scalar value slots
-            // untouched) — the bare flat rc_dec LEAKED every key copy per bind (a
-            // latent leak the map.fold heap-acc loop made observable at a 4MB cap).
-            self.heap_elem_lists.insert(dst);
-        } else if is_heap_elem_list_ty(ty) {
-            self.heap_elem_lists.insert(dst);
-        }
-    }
-
-    /// Extracted from `Self::lower_bind_heap` (pattern-2 uniform-arm split, cog reduction):
-    /// the arm body verbatim, re-narrowed via `let-else`. Pure text move.
-    fn lower_bind_heap_call_computed(&mut self, var: VarId, ty: &Ty, value: &IrExpr) -> Result<(), LowerError> {
-        let IrExprKind::Call { target: CallTarget::Computed { callee }, args, .. } = &value.kind else { unreachable!() };
-        // A tracked closure VAR — or a RECORD-SLOT closure (`h.run("hello")` —
-        // B8's Computed(Member); `closure_block_of_mut` loads the slot borrow).
-        let blk = match self.closure_block_of_mut(callee) {
-            Some(b) => b,
-            None => {
-                return Err(LowerError::Unsupported(
-                    "heap-result record-slot closure call over an unresolvable \
-                     container not in this brick"
-                        .into(),
-                ))
-            }
-        };
-        let repr = repr_of(ty)?;
-        let lowered = self.lower_call_args(args)?;
-        let dst = self.fresh_value();
-        self.emit_closure_call(blk, Some(dst), lowered, Some(repr));
-        self.value_of.insert(var, dst);
-        self.live_heap_handles.push(dst);
-        // A closure-RETURNING closure call (`let triple = make_multiplier(3)` where
-        // `make_multiplier` is a lifted lambda whose tail lifts a capturing lambda):
-        // the result IS a closure block the callee moved out — track it so a later
-        // `triple(4)` dispatches through it (`Op::CallIndirect`) AND its scope-end
-        // drop routes to the recursive `$__drop_closure` (a heap capture would leak
-        // under the default flat rc_dec).
-        if matches!(ty, Ty::Fn { .. }) {
-            self.closure_values.insert(dst);
-        }
-        // The funcref returns its Result/Option in the SAME materialized layout an `ok()`/
-        // `err()` ctor builds (a lifted lambda's body goes through `materialize_result_*`), so
-        // SEED its read-shape — a later `match o { ok/err }` over the bound var then reads its
-        // real tag instead of walling (the higher-order-Result-callback path `fan.map` needs).
-        self.seed_variant_param(dst, ty);
-        // An `Option[List[String]]` closure result (the heap-acc fold's per-iteration
-        // acc): the flat `heap_elem_lists` seed above would free ONE level only,
-        // leaking the inner list's Strings every iteration (a fold loop OOMs) — route
-        // its scope-end drop to the nested `DropListListStr` sweep instead.
-        if crate::lower::is_opt_list_str_ty(ty) {
-            self.heap_elem_lists.remove(&dst);
-            self.list_list_str_lists.insert(dst);
-        }
-        // A MAP closure result (the map.fold heap-acc's per-iteration acc — the
-        // `(a, k, v) => ["fresh": v]` fresh-map closure): the bare
-        // `live_heap_handles` default is a FLAT rc_dec, which frees the map block
-        // but LEAKS its key Strings every iteration (the 100k fold loop OOMs at a
-        // 4MB cap). Route the scope-end drop to the DropListStr sweep — exact for
-        // BOTH map layouts: `Map[String, String]` (interleaved, @4 = 2n, every
-        // slot a String handle) and `Map[String, <scalar>]` (split, @4 = n, the
-        // sweep rc_decs exactly the n key slots; the scalar value slots beyond
-        // are untouched).
-        if matches!(ty, Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Map, a)
-            if a.len() == 2 && matches!(a[0], Ty::String)
-                && (!is_heap_ty(&a[1]) || matches!(a[1], Ty::String)))
-        {
-            self.heap_elem_lists.insert(dst);
-        }
-        Ok(())
-    }
-
-    /// Extracted from `Self::lower_bind_heap` (pattern-2 uniform-arm split, cog reduction):
-    /// the arm body verbatim, re-narrowed via `let-else`. Pure text move.
-    fn lower_bind_heap_match(&mut self, var: VarId, ty: &Ty, value: &IrExpr) -> Result<(), LowerError> {
-        let IrExprKind::Match { subject, arms } = &value.kind else { unreachable!() };
-        // `let e = match <Option[(s1,s2)]> { some(p) => p, none => (f1,f2) }` — the
-        // tuple-unwrap_or desugar output: EXECUTE via component merges + ONE owned
-        // block (no per-arm alloc — cert-clean single object).
-        if let Some(dst) = self.try_lower_scalar_tuple_option_match_bind(subject, arms) {
-            self.value_of.insert(var, dst);
-            self.live_heap_handles.push(dst);
-            self.materialized_aggregates.insert(dst);
-            return Ok(());
-        }
-        // A single-arm tuple-destructure `let offs = match pair { (o, _) => o }` extracting a
-        // HEAP component — semantically `let offs = pair.<i>` (the non-tail tuple-accumulator
-        // `fold` extraction). BORROW the slot handle (the tuple keeps ownership) then ACQUIRE
-        // an OWNED reference (`Op::Dup`, cert `a`) the binding holds + drops at scope end — so
-        // both the tuple's masked drop and this binding's drop are balanced (no double-free, no
-        // leak). Mirrors the `Member`/`TupleIndex` heap-extraction bind arm.
-        if let Some((idx, elem_ty)) = self.tuple_extract_match_index(subject, arms) {
-            if is_heap_ty(&elem_ty) {
-                let synth = Self::synth_tuple_index(subject, idx, elem_ty);
-                if let Some(borrow) = self.try_lower_heap_field_borrow(&synth) {
-                    let dst = self.fresh_value();
-                    self.ops.push(Op::Dup { dst, src: borrow });
-                    self.value_of.insert(var, dst);
-                    self.live_heap_handles.push(dst);
-                    return Ok(());
-                }
-            }
-        }
-        // A TUPLE subject of scalar elements/expressions with a HEAP result
-        // (`let s = match (string.len(x), n % 5) { (2, 0) => "…", _ => "…" }`):
-        // the ordered refinement chain (heap merge), bound + scope-tracked
-        // like any owned heap value.
-        if let Some(dst) = self.try_lower_tuple_refinement_match(subject, arms, ty) {
-            self.value_of.insert(var, dst);
-            if !self.live_heap_handles.contains(&dst) {
-                self.live_heap_handles.push(dst);
-            }
-            return Ok(());
-        }
-        Err(LowerError::Unsupported(
-            "heap-result `match` bound to a let/var cannot be faithfully \
-             computed in this brick (would bind an empty deferred heap value); \
-             the merged result has no sound scope-end drop in the flat certificate"
-                .into(),
-        ))
-    }
-
-    /// Extracted from `Self::lower_bind_heap` (pattern-2 uniform-arm split, cog reduction):
-    /// the arm body verbatim, re-narrowed via `let-else`. Pure text move.
-    fn lower_bind_heap_if(&mut self, var: VarId, ty: &Ty, value: &IrExpr) -> Result<(), LowerError> {
-        let IrExprKind::If { else_, .. } = &value.kind else { unreachable!() };
-        // A VARIANT (Option/Result)-typed `if` RHS (`let $r = if c then ok(T)
-        // else err(e)` — the tail err-raise normalization's two-step bind, and
-        // the hand-written equivalent): EXECUTE via the proven heap-result-if
-        // machinery (each arm materializes + Consumes its value; the merge is
-        // the ONE owned rc=1 block), bind + scope-track it, and SEED its
-        // variant READ-shape so a following `$r!` / `match $r` takes the
-        // executing tag-read path — the same classification a call-result
-        // bind gets (seed_variant_param is read-shape only: no ownership
-        // change, the scope-end drop stays this binding's).
-        if is_variant_ty(ty) {
-            if let IrExprKind::If { cond, then, else_ } = &value.kind {
-                if let Some(obj) = self.try_lower_heap_result_if(cond, then, else_, ty) {
-                    self.value_of.insert(var, obj);
-                    if !self.live_heap_handles.contains(&obj) {
-                        self.live_heap_handles.push(obj);
-                    }
-                    self.seed_variant_param(obj, ty);
-                    return Ok(());
-                }
-            }
-        }
-        // STRAIGHT-LINE identity-else shadow rebind `let acc = if cond then acc + [x] else acc`
-        // (porta `serialize_opts`' 7 stacked optional-arg appends on one `args` slot). The ELSE
-        // arm is EXACTLY the accumulator var — the PROVEN loop-carried `i(id)m` append slot,
-        // UNROLLED straight-line. Drop-old + `SetLocal` the slot in place (the THEN arm only);
-        // the new shadow ALIASES the same slot (NOT re-pushed to live_heap_handles — one
-        // scope-end drop / tail move-out covers it). Each rebind folds to a `(id)` CLoop body
-        // in the certificate (check_line_unroll_sound, the same unit the loop slot proves).
-        if let IrExprKind::Var { id: acc_id } = &else_.kind {
-            if let Some(&acc_local) = self.value_of.get(acc_id) {
-                // The slot must be an OWNED, scope-tracked heap handle (the seed's `[]`/`""`) —
-                // NOT a borrowed param field (`param_values`), whose drop-old would release a
-                // reference we do not own. A borrow falls through to the wall.
-                if self.live_heap_handles.contains(&acc_local)
-                    && !self.param_values.contains(&acc_local)
-                {
-                    let mark = self.ops.len();
-                    if self.try_lower_line_cond_acc(value, *acc_id, acc_local) {
-                        self.value_of.insert(var, acc_local);
-                        return Ok(());
-                    }
-                    self.ops.truncate(mark);
-                }
-            }
-        }
-        Err(LowerError::Unsupported(
-            "heap-result `if` bound to a let/var cannot be faithfully \
-             computed in this brick (would bind an empty deferred heap value); \
-             the merged result has no sound scope-end drop in the flat certificate"
-                .into(),
-        ))
-    }
-
-    /// `let (a, b) = …` — a TUPLE destructuring bind. Two sound shapes:
-    ///
-    /// 1. From a tuple LITERAL `(x, y)` of the same arity — lowered COMPONENT-WISE
-    ///    as ordinary binds (`lower_bind` reused: a `Var` is an alias `Dup`, a
-    ///    literal an `Alloc`/`Const`, a call a real `CallFn` whose caps are
-    ///    captured, NOT elided). The tuple is never materialized.
-    /// 2. From a tracked heap VAR `t` — each HEAP component aliases the WHOLE
-    ///    container `t` (an `Op::Dup`, the container-grain field access of the
-    ///    field-access op), each SCALAR component is a `Const` copy. Aliasing the
-    ///    container keeps it alive for each component's lifetime (a conservative
-    ///    lifetime widening, never a UAF); component-PRECISE identity (`a == t.0`)
-    ///    is deferred to the layout brick.
-    ///
-    /// A `Wildcard` component is ignored. Anything else — a non-tuple/nested/
-    /// constructor/record pattern, or a value that is neither a matching tuple
-    /// literal nor a tracked heap var — stays an explicit `Unsupported` (totality).
-    pub(crate) fn lower_destructure(&mut self, pattern: &IrPattern, value: &IrExpr) -> Result<(), LowerError> {
-        // Shape 1: component-wise from a same-arity tuple LITERAL — each component is
-        // bound to the ACTUAL element (a fresh value / alias, not a container alias),
-        // the most precise lowering. The element's call caps are captured, not elided.
-        if let (IrPattern::Tuple { elements: pats }, IrExprKind::Tuple { elements: vals }) =
-            (pattern, &value.kind)
-        {
-            if pats.len() == vals.len() {
-                for (p, v) in pats.iter().zip(vals) {
-                    match p {
-                        IrPattern::Bind { var, ty } => self.lower_bind(*var, ty, v)?,
-                        IrPattern::Wildcard => {}
-                        // A NESTED tuple sub-pattern `(b, c)` binds against the
-                        // corresponding element value `v` — recurse (the same two sound
-                        // shapes: a same-arity tuple literal binds component-wise, a
-                        // tracked heap var aliases the container).
-                        IrPattern::Tuple { .. } => self.lower_destructure(p, v)?,
-                        _ => {
-                            return Err(LowerError::Unsupported(
-                                "destructure sub-pattern (only a bound var, `_`, or nested tuple) not in this brick"
-                                    .into(),
-                            ))
-                        }
-                    }
-                }
-                return Ok(());
-            }
-        }
-        // Shape 2 (general): materialize/borrow the value as a SUBJECT (a tracked heap
-        // var is borrowed, a fresh heap value is materialized + dropped at scope end),
-        // then bind the pattern CONTAINER-GRAIN (each heap binding aliases the whole
-        // subject — `bind_pattern`). Handles tuple-from-var, constructor, record, and
-        // option/result destructuring; the bound vars drop at scope end.
-        let subject: Option<ValueId> = if is_heap_ty(&value.ty) {
-            match self.lower_call_args(std::slice::from_ref(value))?.into_iter().next() {
-                Some(CallArg::Handle(v)) => Some(v),
-                _ => None,
-            }
-        } else {
-            self.record_elided_calls(value);
-            None
-        };
-        // PRECISE tuple field extraction (the layout brick): a tuple value is a block
-        // [rc][len][cap][f0@12, f1@20, ...]; a destructure (`let (a, b) = t`) loads each field at
-        // its OWN slot instead of the container-grain alias. A SCALAR field is a value COPY; a HEAP
-        // field (`let (inner, z) = n` over `((Int,Int), Int)`) is the BORROWED slot handle (the
-        // tuple keeps ownership through its masked scope-end drop). Without this, `bind_pattern`
-        // aliased the WHOLE container for a heap field and emitted `Const 0` for a scalar field
-        // alongside it = the `8192:2000:0` miscompile.
-        if let IrPattern::Tuple { elements } = pattern {
-            if let Some(subj) = subject {
-                // A CALL-RESULT tuple (`let (v, n) = dispatch(..)`) is a real OWNED block the
-                // callee built (the `lower_tail` Tuple materialize) but `materialized_call_arg`
-                // tracked it only flatly (a plain Drop would LEAK its heap slot, and it is not a
-                // `materialized_aggregate` so the precise destructure below bails to the `Const 0`
-                // container-alias garbage). SEED it as a masked aggregate: record the heap-slot
-                // mask (so the scope-end drop is the recursive `DropListStr` that frees the owned
-                // String/Value slot) + mark it `materialized_aggregates` (so per-slot borrow reads
-                // execute). Only for an owned, still-live result (in `live_heap_handles`) — a
-                // borrowed param/var already carries its own tracking.
-                if !self.materialized_aggregates.contains(&subj)
-                    && self.live_heap_handles.contains(&subj)
-                {
-                    // The tuple's element types: from value.ty when it is a Tuple, ELSE (brick 5) — an
-                    // effect-fn `let (v,p) = f()!` whose `!` Unwrap render_program strips to a Call, so
-                    // value.ty is the effect Result, NOT a Ty::Tuple — from the PATTERN's bound types.
-                    // Without the pattern fallback the seed misses and the destructure container-grains
-                    // (reads slot 0 as the whole handle + slot 1 as Const 0 — the `8212 / 0` garbage).
-                    let elem_tys: Option<Vec<Ty>> = if let Ty::Tuple(tys) = &value.ty {
-                        Some(tys.clone())
-                    } else if matches!(
-                        &value.kind,
-                        IrExprKind::Unwrap { .. } | IrExprKind::Call { .. }
-                    ) {
-                        Some(
-                            elements
-                                .iter()
-                                .map(|p| match p {
-                                    IrPattern::Bind { ty, .. } => ty.clone(),
-                                    _ => Ty::Unit,
-                                })
-                                .collect(),
-                        )
-                    } else {
-                        None
-                    };
-                    if let Some(tys) = elem_tys {
-                        // A (Value, scalar) tuple's Value slot needs the RECURSIVE __drop_value_tuple
-                        // (a flat record_masks rc_dec leaks the Value's nested payload → 10⁴ OOM) — the
-                        // same routing brick 3's construct uses.
-                        let value_tuple = tys.len() == 2
-                            && crate::lower::is_value_ty(&tys[0])
-                            && !is_heap_ty(&tys[1]);
-                        let heap_slots: Vec<usize> =
-                            (0..tys.len()).filter(|&i| is_heap_ty(&tys[i])).collect();
-                        if value_tuple {
-                            self.variant_drop_handles.insert(subj, "value_tuple".to_string());
-                            self.materialized_aggregates.insert(subj);
-                        } else if !heap_slots.is_empty() {
-                            self.record_masks.insert(subj, heap_slots);
-                            self.materialized_aggregates.insert(subj);
-                        }
-                    }
-                }
-                if self.try_lower_tuple_destructure(elements, subj, Some(&value.ty)) {
-                    return Ok(());
-                }
-            }
-        }
-        // PRECISE record field extraction (`let { x, y } = p`) — the record sibling of the tuple
-        // path above. Load each field from its OWN layout slot instead of the container-grain alias
-        // (`bind_pattern` bound every field to the record pointer → `i64.add` on two ptrs / NUL
-        // Strings). A CALL-RESULT record (`let { … } = mk()`) is seeded as a masked aggregate first
-        // (so heap fields borrow + the scope-end drop frees them), exactly like the tuple seed.
-        if let IrPattern::RecordPattern { fields, .. } = pattern {
-            if let Some(subj) = subject {
-                if !self.materialized_aggregates.contains(&subj)
-                    && self.live_heap_handles.contains(&subj)
-                {
-                    if let Some((_, tys)) = self.aggregate_field_tys(&value.ty) {
-                        let heap_slots: Vec<usize> =
-                            (0..tys.len()).filter(|&i| is_heap_ty(&tys[i])).collect();
-                        if !heap_slots.is_empty() {
-                            self.record_masks.insert(subj, heap_slots);
-                        }
-                        self.materialized_aggregates.insert(subj);
-                    }
-                }
-                if self.try_lower_record_destructure(fields, &value.ty, subj) {
-                    return Ok(());
-                }
-            }
-        }
-        self.bind_pattern(pattern, subject)
-    }
 }
